requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beartype==0.16.4
bitarray==2.9.0
boto3==1.34.7
botocore==1.34.7
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
crypto==1.4.1
cytoolz==0.12.2
dnspython==2.4.2
ecdsa==0.18.0
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.7.0
eth-keys==0.4.0
eth-rlp==1.0.0
eth-typing==3.5.2
eth-utils==2.3.1
exceptiongroup==1.2.0
fastapi~=0.104.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
hexbytes==0.3.1
httptools==0.6.1
idna==3.6
importlib-resources==6.1.1
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
loguru==0.7.2
lru-dict==1.2.0
multidict==6.0.4
Naked==0.1.32
packaging==23.2
parsimonious==0.9.0
passlib==1.7.4
pkgutil-resolve-name==1.3.10
pluggy==1.3.0
protobuf==4.25.1
psutil~=5.9.6
pyasn1==0.5.1
pycryptodome==3.19.0
pydantic==1.10.13
pydantic-core==2.14.5
pymongo==4.3.3
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rlp==4.0.0
rpds-py==0.15.2
rsa==4.9
s3transfer==0.10.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy~=2.0.22
starlette==0.27.0
tomli==2.0.1
toolz==0.12.0
typing-extensions==4.9.0
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
web3==6.13.0
websockets==12.0
yarl==1.9.4
zipp==3.17.0
